requirements.txt
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.7
django-extensions==2.2.5
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
gunicorn==20.0.2
idna==2.8
importlib-metadata==0.23
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==7.2.0
netifaces==0.10.9
openapi-codec==1.3.2
packaging==19.2
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.0
pytest-django==3.7.0
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
supervisor==4.1.0
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
